Add stub for loading demo code
I need to figure out the correct way to do this. Maybe this would work
for production build. The problem is that during development the code
gets preprocessed given we depend on it and it hits some cached version
instead of the original. As a result we don't get the code we want...